requirements.txt
# ─── SachBol AI — Full Requirements (local backend) ──────────
# For local development with all ML models.
# For Render deployment, use render_app/requirements.txt instead.

# ── Web framework ─────────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0

# ── LLM APIs ──────────────────────────────────────────────────
groq>=0.9.0

# ── Web scraping ──────────────────────────────────────────────
requests>=2.31.0
cloudscraper>=1.2.71
beautifulsoup4>=4.12.0
pytrends>=4.9.2

# ── News & NLP ────────────────────────────────────────────────
gnews>=0.3.7
textblob>=0.17.1
feedparser>=6.0.10

# ── ML inference (local models) ───────────────────────────────
# CPU-only install (works on any machine, slower):
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
# GPU install (faster, recommended):
#   pip install torch
torch>=2.3.0
transformers>=4.44.0
accelerate>=0.33.0
peft>=0.12.0
sentencepiece>=0.2.0
huggingface_hub>=0.24.0

# ── Training only (uncomment to train) ────────────────────────
# bitsandbytes>=0.43.0
# trl>=0.10.0
# datasets>=2.20.0
